@@ -9,7 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.lang.NonNull;
 
-import javax.websocket.Session;
+import jakarta.websocket.Session;
 
 public class WebsocketSessionAdapter implements SessionAdapter {
 
